+++ src/app/state.rs
@@ -16,6 +16,7 @@ use {
     },
     std::{
         path::{Path, PathBuf},
+        str::FromStr,
     },
     termimad::Area,
 };
@@ -558,3 +559,15 @@ pub trait AppState {
         }
     }
 }
+
+pub fn get_arg<T: Copy + FromStr>(
+    verb_invocation: Option<&VerbInvocation>,
+    internal_exec: &InternalExecution,
+    default: T,
+) -> T {
+    verb_invocation.and_then(|vi| vi.args.as_ref())
+        .or(internal_exec.arg.as_ref())
+        .and_then(|s| s.parse::<T>().ok())
+        .unwrap_or(default)
+}
+