+++ src/syntactic/syntactic_view.rs
@@ -230,28 +230,27 @@ impl SyntacticView {
         false
     }
 
-    pub fn select_previous_line(&mut self) {
-        if let Some(idx) = self.selection_idx {
-            if idx > 0 {
-                self.selection_idx = Some(idx - 1);
-            } else {
-                self.selection_idx = Some(self.lines.len()-1);
-            }
-        } else if !self.lines.is_empty() {
-            self.selection_idx = Some(self.lines.len()-1);
-        }
-        self.ensure_selection_is_visible();
-    }
-
-    pub fn select_next_line(&mut self) {
-        if let Some(idx) = self.selection_idx {
-            if idx < self.lines.len() - 1 {
-                self.selection_idx = Some(idx + 1);
-            } else {
+    pub fn move_selection(&mut self, dy: i32) {
+        if dy > 0 {
+            if let Some(idx) = self.selection_idx {
+                if idx < self.lines.len() - 1 {
+                    self.selection_idx = Some(idx + dy.min(self.lines.len() as i32 - 1 - dy) as usize);
+                } else {
+                    self.selection_idx = Some(0);
+                }
+            } else if !self.lines.is_empty() {
                 self.selection_idx = Some(0);
             }
-        } else if !self.lines.is_empty() {
-            self.selection_idx = Some(0);
+        } else if dy < 0 {
+            if let Some(idx) = self.selection_idx {
+                if idx > 0 {
+                    self.selection_idx = Some(idx - (-dy).min(idx as i32) as usize);
+                } else {
+                    self.selection_idx = Some(self.lines.len()-1);
+                }
+            } else if !self.lines.is_empty() {
+                self.selection_idx = Some(self.lines.len()-1);
+            }
         }
         self.ensure_selection_is_visible();
     }