+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### next
 * dev.log renamed into broot.log
+* `:line_up` and `:line_down` accept an optional count as argument - Fix #301
 
 <a name="v1.0.5"></a>
 ### v1.0.5 - 2020-11-05